refact linux desktop env
