+++ src/platform/linux.rs
@@ -1,5 +1,5 @@
-use super::linux_desktop::{get_desktop_env, Desktop};
 use super::{CursorData, ResultType};
+use desktop::{get_desktop_env, start_check_desktop_env, stop_check_desktop_env, Desktop};
 pub use hbb_common::platform::linux::*;
 use hbb_common::{
     allow_err, bail,
@@ -281,6 +281,7 @@ fn force_stop_server() {
 pub fn start_os_service() {
     stop_rustdesk_servers();
     start_uinput_service();
+    start_check_desktop_env();
 
     let running = Arc::new(AtomicBool::new(true));
     let r = running.clone();
@@ -359,6 +360,8 @@ pub fn start_os_service() {
         }
         sleep_millis(super::SERVICE_INTERVAL);
     }
+
+    stop_check_desktop_env();
 
     if let Some(ps) = user_server.take().as_mut() {
         allow_err!(ps.kill());
@@ -729,4 +732,257 @@ pub fn change_resolution(name: &str, width: usize, height: usize) -> ResultType<
         ])
         .spawn()?;
     Ok(())
+}
+
+mod desktop {
+    use super::{super::SERVICE_INTERVAL, *};
+    use hbb_common::{log, tokio::time};
+    use std::{
+        sync::{
+            atomic::{AtomicBool, Ordering},
+            Arc, Mutex,
+        },
+        time::{Duration, Instant},
+    };
+
+    pub const XFCE4_PANEL: &str = "xfce4-panel";
+    pub const GNOME_SESSION_BINARY: &str = "gnome-session-binary";
+    pub const ENV_DESKTOP_PROTOCAL: &str = "RUSTDESK_PROTOCAL";
+    pub const ENV_DESKTOP_PROTOCAL_WAYLAND: &str = "wayland";
+    pub const ENV_DESKTOP_PROTOCAL__X11: &str = "x11";
+    pub const ENV_DESKTOP_PROTOCAL_UNKNOWN: &str = "unknown";
+
+    lazy_static::lazy_static! {
+        static ref RUNNING: Arc<AtomicBool> = Arc::new(AtomicBool::new(false));
+        static ref DESKTOP_ENV: Arc<Mutex<Desktop>> = Arc::new(Mutex::new(Desktop::new()));
+    }
+
+    #[derive(Copy, Clone, PartialEq, Debug)]
+    pub enum Protocal {
+        Wayland,
+        X11, // Xorg
+        Unknown,
+    }
+
+    impl ToString for Protocal {
+        fn to_string(&self) -> String {
+            match self {
+                Protocal::X11 => ENV_DESKTOP_PROTOCAL__X11.to_owned(),
+                Protocal::Wayland => ENV_DESKTOP_PROTOCAL_WAYLAND.to_owned(),
+                Protocal::Unknown => ENV_DESKTOP_PROTOCAL_UNKNOWN.to_owned(),
+            }
+        }
+    }
+
+    impl From<String> for Protocal {
+        fn from(value: String) -> Self {
+            match &value as &str {
+                ENV_DESKTOP_PROTOCAL__X11 => Protocal::X11,
+                ENV_DESKTOP_PROTOCAL_WAYLAND => Protocal::Wayland,
+                _ => Protocal::Unknown,
+            }
+        }
+    }
+
+    #[derive(Debug, Clone)]
+    pub struct Desktop {
+        pub sid: String,
+        pub protocal: Protocal,
+        pub username: String,
+        pub uid: String,
+        pub display: String,
+        pub xauth: String,
+    }
+
+    fn check_update_env() {
+        let mut desktop = DESKTOP_ENV.lock().unwrap();
+        if !desktop.sid.is_empty() && is_active(&desktop.sid) {
+            return;
+        }
+
+        desktop.refresh();
+        desktop.update_env();
+        log::debug!("desktop env changed, {:?}", &desktop);
+    }
+
+    fn wait_xdesktop(timeout_secs: u64) -> bool {
+        let wait_begin = Instant::now();
+        while wait_begin.elapsed().as_secs() < timeout_secs {
+            let uid = &get_values_of_seat0(&[1])[0];
+            if !uid.is_empty() {
+                return true;
+            }
+
+            if let Ok(output) = run_cmds(format!(
+                "ps -ef | grep -v 'grep' | grep -E 'gnome-session-binary|{}'",
+                XFCE4_PANEL
+            )) {
+                if !output.is_empty() {
+                    log::info!("wait xdesktop: find xclient {}", &output);
+                    return true;
+                }
+            }
+
+            std::thread::sleep(Duration::from_millis(SERVICE_INTERVAL));
+        }
+
+        false
+    }
+
+    pub fn start_check_desktop_env() {
+        std::thread::spawn(|| {
+            if wait_xdesktop(20) {
+                log::info!("Wait desktop: default");
+            } else {
+                log::info!("Wait desktop: none");
+            }
+
+            let interval = time::Duration::from_millis(SERVICE_INTERVAL);
+            RUNNING.store(true, Ordering::SeqCst);
+            while RUNNING.load(Ordering::SeqCst) {
+                check_update_env();
+                std::thread::sleep(interval);
+            }
+            log::info!("xdesktop update thread exit");
+        });
+    }
+
+    pub fn stop_check_desktop_env() {
+        RUNNING.store(false, Ordering::SeqCst);
+    }
+
+    #[inline]
+    pub fn get_desktop_env() -> Desktop {
+        DESKTOP_ENV.lock().unwrap().clone()
+    }
+
+    impl Desktop {
+        pub fn new() -> Self {
+            Self {
+                sid: "".to_owned(),
+                protocal: Protocal::Unknown,
+                username: "".to_owned(),
+                uid: "".to_owned(),
+                display: "".to_owned(),
+                xauth: get_env_var("XAUTHORITY"),
+            }
+        }
+
+        fn update_env(&self) {
+            if self.is_x11() {
+                std::env::set_var("DISPLAY", &self.display);
+                std::env::set_var("XAUTHORITY", &self.xauth);
+                std::env::set_var(ENV_DESKTOP_PROTOCAL, &self.protocal.to_string());
+            } else {
+                std::env::set_var("DISPLAY", "");
+                std::env::set_var("XAUTHORITY", "");
+                std::env::set_var(ENV_DESKTOP_PROTOCAL, &self.protocal.to_string());
+            }
+        }
+
+        #[inline]
+        pub fn is_x11(&self) -> bool {
+            self.protocal == Protocal::X11
+        }
+
+        #[inline]
+        pub fn is_wayland(&self) -> bool {
+            self.protocal == Protocal::Wayland
+        }
+
+        fn get_display(&mut self) {
+            self.display = get_env_tries("DISPLAY", &self.uid, GNOME_SESSION_BINARY, 10);
+            if self.display.is_empty() {
+                self.display = get_env_tries("DISPLAY", &self.uid, XFCE4_PANEL, 10);
+            }
+            if self.display.is_empty() {
+                self.display = Self::get_display_by_user(&self.username);
+            }
+            if self.display.is_empty() {
+                self.display = ":0".to_owned();
+            }
+            self.display = self
+                .display
+                .replace(&whoami::hostname(), "")
+                .replace("localhost", "");
+        }
+
+        fn get_xauth(&mut self) {
+            self.xauth = get_env_tries("XAUTHORITY", &self.uid, GNOME_SESSION_BINARY, 10);
+            if self.xauth.is_empty() {
+                get_env_tries("XAUTHORITY", &self.uid, XFCE4_PANEL, 10);
+            }
+
+            let gdm = format!("/run/user/{}/gdm/Xauthority", self.uid);
+            if self.xauth.is_empty() {
+                self.xauth = if std::path::Path::new(&gdm).exists() {
+                    gdm
+                } else {
+                    let username = &self.username;
+                    if username == "root" {
+                        format!("/{}/.Xauthority", username)
+                    } else {
+                        let tmp = format!("/home/{}/.Xauthority", username);
+                        if std::path::Path::new(&tmp).exists() {
+                            tmp
+                        } else {
+                            format!("/var/lib/{}/.Xauthority", username)
+                        }
+                    }
+                };
+            }
+        }
+
+        fn get_display_by_user(user: &str) -> String {
+            // log::debug!("w {}", &user);
+            if let Ok(output) = std::process::Command::new("w").arg(&user).output() {
+                for line in String::from_utf8_lossy(&output.stdout).lines() {
+                    let mut iter = line.split_whitespace();
+                    let b = iter.nth(2);
+                    if let Some(b) = b {
+                        if b.starts_with(":") {
+                            return b.to_owned();
+                        }
+                    }
+                }
+            }
+            // above not work for gdm user
+            //log::debug!("ls -l /tmp/.X11-unix/");
+            let mut last = "".to_owned();
+            if let Ok(output) = std::process::Command::new("ls")
+                .args(vec!["-l", "/tmp/.X11-unix/"])
+                .output()
+            {
+                for line in String::from_utf8_lossy(&output.stdout).lines() {
+                    let mut iter = line.split_whitespace();
+                    let user_field = iter.nth(2);
+                    if let Some(x) = iter.last() {
+                        if x.starts_with("X") {
+                            last = x.replace("X", ":").to_owned();
+                            if user_field == Some(&user) {
+                                return last;
+                            }
+                        }
+                    }
+                }
+            }
+            last
+        }
+
+        fn refresh(&mut self) {
+            *self = Self::new();
+
+            let seat0_values = get_values_of_seat0(&[0, 1, 2]);
+            if seat0_values[0].is_empty() {
+                return;
+            }
+
+            self.sid = seat0_values[0].clone();
+            self.uid = seat0_values[1].clone();
+            self.username = seat0_values[2].clone();
+            self.protocal = get_display_server_of_session(&self.sid).into();
+            self.get_display();
+            self.get_xauth();
+        }
+    }
 }
